+++ src/RoadCaptain.App.RouteBuilder/ViewModels/DesignTimeMainWindowViewModel.cs
@@ -6,7 +6,6 @@ using System.Collections.Generic;
 using Autofac;
 using Avalonia.Threading;
 using Microsoft.Extensions.Configuration;
-using RoadCaptain.App.RouteBuilder.UseCases;
 using RoadCaptain.Ports;
 using Serilog.Core;
 
@@ -32,6 +31,7 @@ namespace RoadCaptain.App.RouteBuilder.ViewModels
         {
             Route.OutputFilePath = @"C:\git\RoadCaptain\test\RoadCaptain.Tests.Unit\GameState\Repro\Rebel.Route-Italian.Villa.Sprint.Loop.json";
             Route.Load();
+            Route.LoopMode = LoopMode.Infinite;
             Route.Markers.Add(new MarkerViewModel(new Segment(new List<TrackPoint>()){ Id = "test", Name="Test", Type = SegmentType.Climb, Sport = SportType.Cycling}));
             Route.Markers.Add(new MarkerViewModel(new Segment(new List<TrackPoint>()){ Id = "test", Name="Test", Type = SegmentType.Climb, Sport = SportType.Cycling}));
             Route.Markers.Add(new MarkerViewModel(new Segment(new List<TrackPoint>()){ Id = "test", Name="Test", Type = SegmentType.Climb, Sport = SportType.Cycling}));